+++ packages/react-components/react-tag-picker-preview/src/components/TagPicker/useTagPicker.ts
@@ -9,8 +9,7 @@ import type {
 import { optionClassNames } from '@fluentui/react-combobox';
 import { PositioningShorthandValue, resolvePositioningShorthand, usePositioning } from '@fluentui/react-positioning';
 import { useActiveDescendant } from '@fluentui/react-aria';
-import { useComboboxBaseState } from '../../utils/useComboboxBaseState';
-import { ComboboxBaseState } from '../../utils/ComboboxBase.types';
+import { useComboboxBaseState, ComboboxBaseState } from '@fluentui/react-combobox';
 
 /**
  * Create the state required to render Picker.
@@ -22,9 +21,10 @@ import { ComboboxBaseState } from '../../utils/ComboboxBase.types';
  */
 export const useTagPicker_unstable = (props: TagPickerProps): TagPickerState => {
   const popoverId = useId('picker-listbox');
-  const triggerInnerRef = React.useRef<HTMLInputElement>(null);
+  const triggerInnerRef = React.useRef<HTMLInputElement | HTMLButtonElement>(null);
   const secondaryActionRef = React.useRef<HTMLSpanElement>(null);
-  const { positioning, size = 'medium', disabled = false } = props;
+  const tagPickerGroupRef = React.useRef<HTMLDivElement>(null);
+  const { positioning, size = 'medium', inline = false } = props;
 
   // Set a default set of fallback positions to try if the dropdown does not fit on screen
   const fallbackPositions: PositioningShorthandValue[] = ['above', 'after', 'after-top', 'before', 'before-top'];
@@ -72,10 +72,12 @@ export const useTagPicker_unstable = (props: TagPickerProps): TagPickerState => 
 
   const setOpen: ComboboxBaseState['setOpen'] = useEventCallback((event, newValue) => {
     // if event comes from secondary action, ignore it
-    if (isHTMLElement(event.target) && elementContains(secondaryActionRef.current, event.target)) {
-      return;
-    }
-    if (disabled) {
+    // if event comes from tags, ignore it
+    if (
+      isHTMLElement(event.target) &&
+      (elementContains(secondaryActionRef.current, event.target) ||
+        elementContains(tagPickerGroupRef.current, event.target))
+    ) {
       return;
     }
     comboboxState.setOpen(event, newValue);
@@ -87,12 +89,14 @@ export const useTagPicker_unstable = (props: TagPickerProps): TagPickerState => 
     trigger,
     popover: comboboxState.open || comboboxState.hasFocus ? popover : undefined,
     popoverId,
-    disabled,
+    disabled: comboboxState.disabled,
     triggerRef: useMergedRefs(triggerInnerRef, activeParentRef),
     popoverRef: useMergedRefs(listboxRef, containerRef),
     secondaryActionRef,
+    tagPickerGroupRef,
     targetRef,
     size,
+    inline,
     open: comboboxState.open,
     mountNode: comboboxState.mountNode,
     onOptionClick: useEventCallback(event => {
@@ -110,6 +114,7 @@ export const useTagPicker_unstable = (props: TagPickerProps): TagPickerState => 
     setValue: comboboxState.setValue,
     multiselect: comboboxState.multiselect,
     value: comboboxState.value,
+    freeform: comboboxState.freeform,
   };
 };