+++ packages/react-components/react-tag-picker-preview/src/components/TagPickerInput/useTagPickerInput.tsx
@@ -3,9 +3,16 @@ import * as ReactDOM from 'react-dom';
 import type { TagPickerInputProps, TagPickerInputState } from './TagPickerInput.types';
 import { useActiveDescendantContext } from '@fluentui/react-aria';
 import { useTagPickerContext_unstable } from '../../contexts/TagPickerContext';
-import { useMergedRefs, getIntrinsicElementProps, useEventCallback } from '@fluentui/react-utilities';
-import { useInputTriggerSlot } from '../../utils/useInputTriggerSlot';
+import {
+  useMergedRefs,
+  getIntrinsicElementProps,
+  useEventCallback,
+  useIsomorphicLayoutEffect,
+} from '@fluentui/react-utilities';
 import { Backspace, Enter } from '@fluentui/keyboard-keys';
+import { useInputTriggerSlot } from '@fluentui/react-combobox';
+import { useFieldControlProps_unstable } from '@fluentui/react-field';
+import { tagPickerInputCSSRules } from '../../utils/tokens';
 
 /**
  * Create the state required to render TagPickerInput.
@@ -20,8 +27,10 @@ export const useTagPickerInput_unstable = (
   props: TagPickerInputProps,
   ref: React.Ref<HTMLInputElement>,
 ): TagPickerInputState => {
+  props = useFieldControlProps_unstable(props, { supportsLabelFor: true, supportsRequired: true, supportsSize: true });
   const { controller: activeDescendantController } = useActiveDescendantContext();
   const size = useTagPickerContext_unstable(ctx => ctx.size);
+  const freeform = useTagPickerContext_unstable(ctx => ctx.freeform);
   const contextDisabled = useTagPickerContext_unstable(ctx => ctx.disabled);
   const {
     triggerRef,
@@ -36,7 +45,27 @@ export const useTagPickerInput_unstable = (
     multiselect,
     popoverId,
     value: contextValue,
-  } = usePickerContext();
+  } = useTagPickerContexts();
+
+  useIsomorphicLayoutEffect(() => {
+    if (!triggerRef.current) {
+      return;
+    }
+    setTagPickerInputStretchStyle(triggerRef.current);
+  }, [selectedOptions, triggerRef]);
+
+  useIsomorphicLayoutEffect(() => {
+    if (triggerRef.current) {
+      const input = triggerRef.current;
+      const cb = () => {
+        setTagPickerInputStretchStyle(input);
+      };
+      input.addEventListener('input', cb);
+      return () => {
+        input.removeEventListener('input', cb);
+      };
+    }
+  }, [triggerRef]);
 
   const { value = contextValue, disabled = contextDisabled } = props;
 
@@ -50,10 +79,14 @@ export const useTagPickerInput_unstable = (
       onKeyDown: useEventCallback((event: React.KeyboardEvent<HTMLInputElement>) => {
         props.onKeyDown?.(event);
         if (event.key === Enter) {
-          ReactDOM.unstable_batchedUpdates(() => {
-            setValue(undefined);
-            setOpen(event, false);
-          });
+          if (open) {
+            ReactDOM.unstable_batchedUpdates(() => {
+              setValue(undefined);
+              setOpen(event, false);
+            });
+          } else {
+            setOpen(event, true);
+          }
         }
 
         if (event.key === Backspace && value?.length === 0 && selectedOptions.length) {
@@ -71,7 +104,7 @@ export const useTagPickerInput_unstable = (
     useMergedRefs(triggerRef, ref),
     {
       activeDescendantController,
-      freeform: props.freeform,
+      freeform,
       state: {
         clearSelection,
         getOptionById,
@@ -99,9 +132,9 @@ export const useTagPickerInput_unstable = (
   return state;
 };
 
-function usePickerContext() {
+function useTagPickerContexts() {
   return {
-    triggerRef: useTagPickerContext_unstable(ctx => ctx.triggerRef),
+    triggerRef: useTagPickerContext_unstable(ctx => ctx.triggerRef) as React.RefObject<HTMLInputElement>,
     clearSelection: useTagPickerContext_unstable(ctx => ctx.clearSelection),
     getOptionById: useTagPickerContext_unstable(ctx => ctx.getOptionById),
     open: useTagPickerContext_unstable(ctx => ctx.open),
@@ -115,3 +148,25 @@ function usePickerContext() {
     popoverId: useTagPickerContext_unstable(ctx => ctx.popoverId),
   };
 }
+
+/**
+ * while typing the user might need a bit more of space to see the text,
+ * which means the input should stretch to 100% width
+ * occupying a whole new line.
+ *
+ * This function will set the CSS variable `--width` to `100%` if the scrollWidth is greater than the offsetWidth,
+ * meaning the text is overflowing the input.
+ *
+ * @param input - input element to apply the style
+ * @returns void
+ */
+const setTagPickerInputStretchStyle = (input: HTMLInputElement) => {
+  // first we need to remove the CSS variable
+  // to properly calculate the difference between scrollWidth and offsetWidth
+  input.style.removeProperty(tagPickerInputCSSRules.width);
+  if (input.scrollWidth > input.offsetWidth + 1) {
+    input.style.setProperty(tagPickerInputCSSRules.width, '100%');
+  } else {
+    input.style.removeProperty(tagPickerInputCSSRules.width);
+  }
+};