+++ packages/react-components/react-tag-picker-preview/stories/TagPicker/index.stories.tsx
@@ -19,6 +19,9 @@ export { Appearance } from './TagPickerAppearance.stories';
 export { Disabled } from './TagPickerDisabled.stories';
 export { ExpandIcon } from './TagPickerExpandIcon.stories';
 export { SecondaryAction } from './TagPickerSecondaryAction.stories';
+export { Grouped } from './TagPickerGrouped.stories';
+export { Freeform } from './TagPickerFreeform.stories';
+export { InteractionTagExample as InteractionTag } from './TagPickerInteractionTag.stories';
 
 export default {
   title: 'Preview Components/Tag Picker',